Introduce Github::OAuthClient and refactor shortcut Github::Client
